Revert "Add utils function to convert from size_t to int32_t."
This reverts commit ff79e6ac7782e9e04aa298c3848c3c6a6a56cd2a.

+++ src/shogun/lib/parameter_observers/ParameterObserverCV.h
@@ -66,7 +66,7 @@ namespace shogun
 		 * by this observer.
 		 * @return number of runs.
 		 */
-		const int32_t get_num_observations() const;
+		const size_t get_num_observations() const;
 
 		/**
 		 * Get a CrossValidationStorage object which will store
